First Phase (Initial Model training)
Wrote and trained an AI Handwritten Digits Recognition model. 
The successfull prediciton score is 96-98%.
Requires an upload of 28x28 pixel grayscale digit image to be able to process.
Next step to make it accept normal colored photos of users taking a picture of the handwritten digit and uploading to the model.
Further stage to make a simple web interface and deploy it.